:bath: Move and clean up
