+++ src/app/routes/getInitialData/utils/getBaseUrl.js
